requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pymongo==4.7.2
qdrant-client==1.8.0
openai==1.30.1
langchain==0.1.20
langchain-community==0.0.37
langchain-openai==0.1.6
langchain-neo4j==0.0.7
neo4j==5.19.0
numpy==1.26.4
python-dotenv==1.0.1
httpx==0.27.0
supervisor==4.2.5
selenium==4.15.2
webdriver-manager==4.0.1
schedule==1.2.1
pydantic==2.5.2
langchain-experimental==0.0.45
langchain-core==0.1.17
beautifulsoup4==4.12.2